+++ packages/amplify-codegen-e2e-tests/src/codegen-tests-base/graphql-codegen.ts
@@ -6,7 +6,7 @@ import {
     AmplifyFrontendConfig,
     generateStatementsAndTypes,
     createRandomName
-} from "amplify-codegen-e2e-core";
+} from "@aws-amplify/amplify-codegen-e2e-core";
 import { existsSync } from "fs";
 import path from 'path';
 import { isNotEmptyDir, generateSourceCode } from '../utils';