+++ packages/amplify-codegen-e2e-tests/src/codegen-tests-base/push-codegen.ts
@@ -7,7 +7,7 @@ import {
     AmplifyFrontendConfig,
     amplifyPushWithCodegenUpdate,
     updateAPIWithResolutionStrategyWithModels
-} from "amplify-codegen-e2e-core";
+} from "@aws-amplify/amplify-codegen-e2e-core";
 import { existsSync } from "fs";
 import path from 'path';
 import { isNotEmptyDir } from '../utils';