+++ packages/amplify-codegen-e2e-tests/src/schema-api-directives/tests/function-example2.ts
@@ -7,7 +7,7 @@ import {
   addApiWithCognitoUserPoolAuthTypeWhenAuthExists,
   updateAuthAddUserGroups,
   addAuthWithDefault,
-} from 'amplify-codegen-e2e-core';
+} from '@aws-amplify/amplify-codegen-e2e-core';
 
 import { updateFunctionNameInSchema } from '../functionTester';