+++ packages/amplify-codegen/src/utils/index.js
@@ -15,9 +15,6 @@ const isCodegenConfigured = require('./isCodegenConfigured');
 const getSDLSchemaLocation = require('./getSDLSchemaLocation');
 const switchToSDLSchema = require('./switchToSDLSchema');
 const ensureIntrospectionSchema = require('./ensureIntrospectionSchema');
-const getModelgenPackage = require('./getModelgenPackage');
-const getDocsgenPackage = require('./getDocsgenPackage');
-const getTypesgenPackage = require('./getTypesgenPackage');
 
 module.exports = {
   getAppSyncAPIDetails,
@@ -37,7 +34,4 @@ module.exports = {
   getSDLSchemaLocation,
   switchToSDLSchema,
   ensureIntrospectionSchema,
-  getModelgenPackage,
-  getDocsgenPackage,
-  getTypesgenPackage
 };